[Refactor] Data 패키지 추가 및 미사용 클래스 제거
