+++ app/src/main/java/com/and04/naturealbum/data/room/Data.kt
@@ -1,4 +1,4 @@
-package com.and04.naturealbum
+package com.and04.naturealbum.data.room
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
